+++ bitframe-utils/presenters/core/src/commonTest/kotlin/changes/ChangeRemarkBuilderTest.kt
@@ -26,7 +26,6 @@ class ChangeRemarkBuilderTest {
             previous = 100.TZS,
             current = 300.USD
         )
-//        expect(change).toBe<ChangeRemark.Indeterminate<Money>>()
         expect(change).toBe<ChangeRemark.Indeterminate>()
     }