+++ pimonitor/pimonitor-core/src/commonMain/kotlin/pimonitor/core/investments/BusinessInvestmentsServiceCore.kt
@@ -6,15 +6,15 @@ import bitframe.core.RequestBody
 import kotlinx.collections.interoperable.List
 import later.Later
 import pimonitor.core.investments.params.InvestmentDisbursementParams
-import pimonitor.core.investments.params.InvestmentsParams
-import pimonitor.core.business.utils.disbursements.Disbursement
+import pimonitor.core.investments.params.InvestmentParams
+import pimonitor.core.utils.disbursements.Disbursement
 import kotlin.js.JsExport
 import kotlin.js.JsName
 
 @JsExport
 interface BusinessInvestmentsServiceCore {
     @JsName("_ignore_capture")
-    fun capture(rb: RequestBody.Authorized<InvestmentsParams>): Later<Investment>
+    fun capture(rb: RequestBody.Authorized<InvestmentParams>): Later<Investment>
 
     /**
      * @param rb takes in a [RequestBody.Authorized] of a businessId