+++ pimonitor/pimonitor-sdk/client/core/src/commonMain/kotlin/pimonitor/client/business/utils/disbursements/DisbursementFormParams.kt
@@ -1,6 +1,0 @@
-package pimonitor.client.business.utils.disbursements
-
-data class DisbursementFormParams(
-    override val amount: String,
-    override val date: String
-) : DisbursementRawFormParams