+++ pimonitor/pimonitor-sdk/client/core/src/commonMain/kotlin/pimonitor/client/investments/InvestmentIntent.kt
@@ -1,22 +1,22 @@
 package pimonitor.client.investments
 
-import pimonitor.client.business.utils.disbursements.DisbursementRawFormParams
 import pimonitor.core.businesses.MonitoredBusinessBasicInfo
 import pimonitor.core.investments.InvestmentSummary
-import pimonitor.core.investments.params.InvestmentsRawParams
+import pimonitor.core.investments.params.InvestmentRawParams
+import pimonitor.core.utils.disbursements.params.DisbursementRawParams
 import presenters.table.Row
 
 sealed class InvestmentIntent {
-    object LoadAllInvestments : InvestmentIntent()
+    data class LoadAllInvestments(val businessId: String?) : InvestmentIntent()
 
-    data class ShowCreateInvestmentForm(val business: MonitoredBusinessBasicInfo?, val params: InvestmentsRawParams?) : InvestmentIntent()
-    data class SendCreateInvestmentForm(val params: InvestmentsRawParams) : InvestmentIntent()
+    data class ShowCreateInvestmentForm(val business: MonitoredBusinessBasicInfo?, val params: InvestmentRawParams?) : InvestmentIntent()
+    data class SendCreateInvestmentForm(val params: InvestmentRawParams) : InvestmentIntent()
 
-    data class ShowUpdateInvestmentForm(val investment: InvestmentSummary, val params: InvestmentsRawParams?) : InvestmentIntent()
-    data class SendUpdateInvestmentForm(val investment: InvestmentSummary, val params: InvestmentsRawParams) : InvestmentIntent()
+    data class ShowUpdateInvestmentForm(val investment: InvestmentSummary, val params: InvestmentRawParams?) : InvestmentIntent()
+    data class SendUpdateInvestmentForm(val investment: InvestmentSummary, val params: InvestmentRawParams) : InvestmentIntent()
 
-    data class ShowDisbursementForm(val investment: InvestmentSummary, val params: DisbursementRawFormParams?) : InvestmentIntent()
-    data class SendDisbursementForm(val investment: InvestmentSummary, val params: DisbursementRawFormParams) : InvestmentIntent()
+    data class ShowDisbursementForm(val investment: InvestmentSummary, val params: DisbursementRawParams?) : InvestmentIntent()
+    data class SendDisbursementForm(val investment: InvestmentSummary, val params: DisbursementRawParams) : InvestmentIntent()
 
     data class ShowDeleteOneInvestmentDialog(val investment: InvestmentSummary) : InvestmentIntent()
     data class SendDeleteOneInvestmentIntent(val investment: InvestmentSummary) : InvestmentIntent()