+++ pimonitor/pimonitor-sdk/client/core/src/commonMain/kotlin/pimonitor/client/utils/disbursements/forms/UpdateDisbursementForm.kt
@@ -1,17 +1,17 @@
-package pimonitor.client.business.utils.disbursements
+package pimonitor.client.utils.disbursements.forms
 
 import identifier.Name
-import pimonitor.core.business.utils.disbursements.Disbursement
+import pimonitor.core.utils.disbursements.Disbursement
 import presenters.forms.Form
 import presenters.forms.FormActionsBuildingBlock
-import pimonitor.client.business.utils.disbursements.DisbursementFields as Fields
-import pimonitor.client.business.utils.disbursements.DisbursementRawFormParams as Params
+import pimonitor.client.utils.disbursements.params.DisbursementFields as Fields
+import pimonitor.core.utils.disbursements.params.DisbursementRawParams as Params
 
-class EditDisbursementForm(
+fun UpdateDisbursementForm(
     disbursement: Disbursement,
     params: Params? = null,
     block: FormActionsBuildingBlock<Params>
-) : Form<Fields, Params> by Form(
+) = Form(
     heading = "Edit Disbursement",
     details = "Edit ${Name(disbursement.by.name).first}'s ${disbursement.amount} disbursement",
     fields = Fields(params, disbursement),